refactor: finishedGames refactor

+++ backend/routes/api/finishedGames.js
@@ -9,41 +9,40 @@ router.post("/finish-game", auth.required, (req, res, next) => {
 	const { payload: { id } } = req;
 	const { body: { game } } = req;
 
-	return Users.findById(id)
-		.then((user) => {
-			if(!user) {
-				return res.json({ user: "Access is denied" });
+	return Users.findById(id, (err, user) => {
+		if(!user || err) {
+			return res.json({ user: "Access is denied" });
+		}
+		//access is allowed
+		OpenedGame.findByIdAndDelete(game.id, async (err, gameFound) => {
+			if (err){
+				console.log(err);  
 			}
-			//access is allowed
-			OpenedGame.findByIdAndDelete(game.id, async (err, gameFound) => {
-				if (err){
-					console.log(err);  
-				} else {
-					const newGame = new FinishedGame();
-					newGame.players = gameFound.players;
-					newGame.moves = gameFound.moves;
-					newGame.timeToGo = gameFound.timeToGo;
-					newGame.timeWhite = game.timeWhite;
-					newGame.timeBlack = game.timeBlack;
-					newGame.winner = game.winner;
+			const newGame = new FinishedGame();
+			newGame.players = gameFound.players;
+			newGame.moves = gameFound.moves;
+			newGame.timeToGo = gameFound.timeToGo;
+			newGame.timeWhite = game.timeWhite;
+			newGame.timeBlack = game.timeBlack;
+			newGame.winner = game.winner;
 
-					newGame.save().then((gameSaved) => {
-						Users.findOne({_id: gameSaved.players.player1ID}).then(userFound => {
-							userFound.activeGame = "";
-							userFound.rating = game.finalElo1;
-							userFound.save();
-						});
-						Users.findOne({_id: gameSaved.players.player2ID}).then(userFound => {
-							userFound.activeGame = "";
-							userFound.rating = game.finalElo2;
-							userFound.save();
-						});
+			let gameSaved = await newGame.save();
 
-						res.json({ game: gameSaved.toJSON() });
-					});
-				}
+			Users.findOne({_id: gameSaved.players.player1ID}).then(userFound => {
+				userFound.activeGame = "";
+				userFound.rating = game.finalElo1;
+				userFound.save();
 			});
+
+			Users.findOne({_id: gameSaved.players.player2ID}).then(userFound => {
+				userFound.activeGame = "";
+				userFound.rating = game.finalElo2;
+				userFound.save();
+			});
+
+			res.json({ game: gameSaved.toJSON() });
 		});
+	});			
 });
 
 router.post("/get-finish-game", auth.required, (req, res, next) => {
